+++ src/index/updater/inscription_updater.rs
@@ -24,6 +24,7 @@ enum Origin {
 pub(super) struct InscriptionUpdater<'a, 'db, 'tx> {
   flotsam: Vec<Flotsam>,
   height: u64,
+  height_to_inscription_id: &'a mut MultimapTable<'db, 'tx, u64, &'static InscriptionIdValue>,
   id_to_children:
     &'a mut MultimapTable<'db, 'tx, &'static InscriptionIdValue, &'static InscriptionIdValue>,
   id_to_satpoint: &'a mut Table<'db, 'tx, &'static InscriptionIdValue, &'static SatPointValue>,
@@ -48,6 +49,7 @@ pub(super) struct InscriptionUpdater<'a, 'db, 'tx> {
 impl<'a, 'db, 'tx> InscriptionUpdater<'a, 'db, 'tx> {
   pub(super) fn new(
     height: u64,
+    height_to_inscription_id: &'a mut MultimapTable<'db, 'tx, u64, &'static InscriptionIdValue>,
     id_to_children: &'a mut MultimapTable<
       'db,
       'tx,
@@ -84,6 +86,7 @@ impl<'a, 'db, 'tx> InscriptionUpdater<'a, 'db, 'tx> {
     Ok(Self {
       flotsam: Vec::new(),
       height,
+      height_to_inscription_id,
       id_to_children,
       id_to_satpoint,
       value_receiver,
@@ -194,7 +197,7 @@ impl<'a, 'db, 'tx> InscriptionUpdater<'a, 'db, 'tx> {
 
           let sat = Self::calculate_sat(input_sat_ranges, offset);
 
-          log::info!("processing reinscription {inscription_id} on sat {:?}: sequence number {seq_num}, inscribed offsets {:?}", sat, inscribed_offsets);
+          log::debug!("processing reinscription {inscription_id} on sat {:?}: sequence number {seq_num}, inscribed offsets {:?}", sat, inscribed_offsets);
 
           Some(Curse::Reinscription)
         } else {
@@ -202,7 +205,7 @@ impl<'a, 'db, 'tx> InscriptionUpdater<'a, 'db, 'tx> {
         };
 
         if curse.is_some() {
-          log::info!("found cursed inscription {inscription_id}: {:?}", curse);
+          log::debug!("found cursed inscription {inscription_id}: {:?}", curse);
         }
 
         let cursed = if let Some(Curse::Reinscription) = curse {
@@ -224,7 +227,7 @@ impl<'a, 'db, 'tx> InscriptionUpdater<'a, 'db, 'tx> {
             })
             .unwrap_or(false);
 
-          log::info!("{inscription_id}: is first reinscription: {first_reinscription}, initial inscription is cursed: {initial_inscription_is_cursed}");
+          log::debug!("{inscription_id}: is first reinscription: {first_reinscription}, initial inscription is cursed: {initial_inscription_is_cursed}");
 
           !(initial_inscription_is_cursed && first_reinscription)
         } else {
@@ -234,7 +237,7 @@ impl<'a, 'db, 'tx> InscriptionUpdater<'a, 'db, 'tx> {
         let unbound = current_input_value == 0 || curse == Some(Curse::UnrecognizedEvenField);
 
         if curse.is_some() || unbound {
-          log::info!(
+          log::debug!(
             "indexing inscription {inscription_id} with curse {:?} as cursed {} and unbound {}",
             curse,
             cursed,
@@ -435,6 +438,9 @@ impl<'a, 'db, 'tx> InscriptionUpdater<'a, 'db, 'tx> {
     let inscription_id = flotsam.inscription_id.store();
     let unbound = match flotsam.origin {
       Origin::Old { old_satpoint } => {
+        self
+          .height_to_inscription_id
+          .insert(&self.height, &inscription_id)?;
         self.satpoint_to_id.remove_all(&old_satpoint.store())?;
 
         false