+++ src/options.rs
@@ -49,6 +49,8 @@ pub(crate) struct Options {
   pub(crate) index_runes_pre_alpha_i_agree_to_get_rekt: bool,
   #[arg(long, help = "Track location of all satoshis.")]
   pub(crate) index_sats: bool,
+  #[arg(long, help = "Inhibit the display of the progress bar while updating the index.")]
+  pub(crate) no_progress_bar: bool,
   #[arg(long, short, help = "Use regtest. Equivalent to `--chain regtest`.")]
   pub(crate) regtest: bool,
   #[arg(long, help = "Connect to Bitcoin Core RPC at <RPC_URL>.")]
@@ -59,6 +61,8 @@ pub(crate) struct Options {
   pub(crate) testnet: bool,
   #[arg(long, default_value = "ord", help = "Use wallet named <WALLET>.")]
   pub(crate) wallet: String,
+  #[arg(long, short, help = "Don't check for standard wallet descriptors.")]
+  pub(crate) ignore_descriptors: bool,
 }
 
 impl Options {
@@ -258,6 +262,7 @@ impl Options {
         client.load_wallet(&self.wallet)?;
       }
 
+      if !self.ignore_descriptors {
       let descriptors = client.list_descriptors(None)?.descriptors;
 
       let tr = descriptors
@@ -272,6 +277,7 @@ impl Options {
 
       if tr != 2 || descriptors.len() != 2 + rawtr {
         bail!("wallet \"{}\" contains unexpected output descriptors, and does not appear to be an `ord` wallet, create a new wallet with `ord wallet create`", self.wallet);
+      }
       }
     }