+++ src/subcommand.rs
@@ -13,6 +13,7 @@ pub mod subsidy;
 pub mod supply;
 pub mod teleburn;
 pub mod traits;
+pub mod transfer;
 pub mod wallet;
 
 #[derive(Debug, Parser)]
@@ -43,6 +44,8 @@ pub(crate) enum Subcommand {
   Teleburn(teleburn::Teleburn),
   #[command(about = "Display satoshi traits")]
   Traits(traits::Traits),
+  #[command(about = "Modify transfer log table")]
+  Transfer(transfer::Transfer),
   #[command(subcommand, about = "Wallet commands")]
   Wallet(wallet::Wallet),
 }
@@ -68,6 +71,7 @@ impl Subcommand {
       Self::Supply => supply::run(),
       Self::Teleburn(teleburn) => teleburn.run(),
       Self::Traits(traits) => traits.run(),
+      Self::Transfer(transfer) => transfer.run(options),
       Self::Wallet(wallet) => wallet.run(options),
     }
   }