+++ src/subcommand/wallet/send.rs
@@ -4,6 +4,16 @@ use {super::*, crate::subcommand::wallet::transaction_builder::Target, crate::wa
 pub(crate) struct Send {
   address: Address<NetworkUnchecked>,
   outgoing: Outgoing,
+  #[arg(
+    long,
+    help = "Consider spending outpoint <UTXO>, even if it is unconfirmed or contains inscriptions"
+  )]
+  utxo: Vec<OutPoint>,
+  #[clap(
+    long,
+    help = "Only spend outpoints given with --utxo when sending inscriptions or satpoints"
+  )]
+  pub(crate) coin_control: bool,
   #[arg(long, help = "Use fee rate of <FEE_RATE> sats/vB")]
   fee_rate: FeeRate,
   #[arg(
@@ -32,7 +42,20 @@ impl Send {
 
     let client = options.bitcoin_rpc_client_for_wallet_command(false)?;
 
-    let unspent_outputs = index.get_unspent_outputs(Wallet::load(&options)?)?;
+    let mut unspent_outputs = if self.coin_control {
+      BTreeMap::new()
+    } else {
+      index.get_unspent_outputs(Wallet::load(&options)?)?
+    };
+
+    for outpoint in &self.utxo {
+      unspent_outputs.insert(
+        *outpoint,
+        Amount::from_sat(
+          client.get_raw_transaction(&outpoint.txid, None)?.output[outpoint.vout as usize].value,
+        ),
+      );
+    }
 
     let locked_outputs = index.get_locked_outputs(Wallet::load(&options)?)?;
 
@@ -51,6 +74,9 @@ impl Send {
         .get_inscription_satpoint_by_id(id)?
         .ok_or_else(|| anyhow!("Inscription {id} not found"))?,
       Outgoing::Amount(amount) => {
+        if self.coin_control || !self.utxo.is_empty() {
+          bail!("--coin_control and --utxo don't work when sending cardinals");
+        }
         Self::lock_inscriptions(&client, inscriptions, unspent_outputs)?;
         let txid = Self::send_amount(&client, amount, address, self.fee_rate.n())?;
         return Ok(Box::new(Output { transaction: txid }));